Added ScreenFade class and set up player to play spawn effect
Updated PlayerSpawnUI to inherit from a new ScreenFadeEffectClass
Updated PlayerStatus to activate the spawn effect
