+++ Assets/Menus/Scripts/ScreenFadeEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8cc8c092ca83e1b4aa395caa5e2505bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
